Post Migration

+++ app/Http/Controllers/ThemeController.php
@@ -15,6 +15,7 @@ class ThemeController extends Controller
      */
     public function index()
     {
+        $this->authorize('view', Theme::class);
         $themes = Theme::paginate(15);
         return view('administration.themes.index', compact('themes'));
     }
@@ -26,7 +27,7 @@ class ThemeController extends Controller
      */
     public function create()
     {
-        //
+        $this->authorize('create', Theme::class);
     }
 
     /**
@@ -48,6 +49,7 @@ class ThemeController extends Controller
      */
     public function show(Theme $theme)
     {
+        $this->authorize('view', Theme::class);
         return view('administration.themes.show', compact('theme'));
     }
 
@@ -59,7 +61,7 @@ class ThemeController extends Controller
      */
     public function edit($id)
     {
-        //
+        $this->authorize('view', Theme::class);
     }
 
     /**
@@ -71,6 +73,7 @@ class ThemeController extends Controller
      */
     public function update(Request $request, Theme $theme)
     {
+        $this->authorize('update', Theme::class);
         Theme::where('is_default', '=', 1)->update(['is_default' => 0]);
         $theme->update(['is_default' => 1, 'last_modified_by' => Auth::id()]);
         return redirect()->route('themes.show', compact('theme'))->withSuccess("Theme $theme->name was set as default");
@@ -84,6 +87,7 @@ class ThemeController extends Controller
      */
     public function destroy(Theme $theme)
     {
+        $this->authorize('delete', Theme::class);
         if(!$theme->is_default){
             $theme->update(['deleted_by'=> Auth::id()]);
             $theme->delete();
